Split CalItemChanges into Balance and Changes files.
